impl st88 numeric parser
Wrote a DFA as per the bluebook specs, and added a little function that
executes it and if it matches, stores the token parsed as literal. It
works but the way to detect if the numeric dfa should be executed is a
bit wanky... should figure out a better, more extensible way.