move the syms () and # into individual tokens
Currently facing an issue with parsing that lead to this decision.
The problem resides on the special syntatic rules that apply when inside
an array.

An array is opened w/ the symbols #(, and closed with ). Arrays can have
nested arrays inside, but those don't need to be preceded with the pound
symbol, e.g. #( ( 1 2 3 ) ( 4 5 6 ) ), an array of two arrays.

Arrays can store symbols too, but as with nested arrays, symbols must
not be preceded with the pound symbol. #( abc cde ) is an array of 2
symbols.

Thus, there are two ways of constructing an array and a symbol,
depending if we are inside an array or not. This is why parsing #( as
<ARRAY_OPEN> but ( as <PAR_OPEN> doesn't seem like the correct thing to
do. Same as with parsing #A as <SYMBOL> but A as <VAR>. So I thought it
may be a better idea to parse each character individually and then verify
its syntatic and semantical veracity later.

Tho, with my lack of experience, I don't know where this decision will
lead to, hope it ain't too bad :)